[#127] Move SplashUiState into separate file
